+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.6] - 2026-07-24
+
+### Changed
+- 🏷️ Hide climate/fan device name when only one entity is controlled
+- 🎨 Remove gray background from climate/fan control area
+
 ## [1.2.5] - 2026-07-24
 
 ### Changed